refactor: organiza funções do SignUpForm em hooks
